+++ LandonHotel/Services/IBookingService.cs
@@ -5,6 +5,6 @@ namespace LandonHotel.Services
     public interface IBookingService
     {
         bool IsBookingValid(int roomId, Booking booking);
-        int CalculateBookingCost(int roomId, Booking booking);
+        double CalculateBookingCost(int roomId, Booking booking);
     }
 }
